+++ src/main/java/com/github/yunabraska/githubworkflow/model/LocalReferenceResolver.java
@@ -1,37 +1,28 @@
 package com.github.yunabraska.githubworkflow.model;
 
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementResolveResult;
-import com.intellij.psi.PsiManager;
 import com.intellij.psi.PsiPolyVariantReference;
 import com.intellij.psi.PsiReferenceBase;
 import com.intellij.psi.ResolveResult;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import static com.github.yunabraska.githubworkflow.helper.PsiElementHelper.getProject;
-import static com.github.yunabraska.githubworkflow.services.ActionReferenceContributor.ACTION_KEY;
-import static java.util.Optional.ofNullable;
-
 public class LocalReferenceResolver extends PsiReferenceBase<PsiElement> implements PsiPolyVariantReference {
 
-    public LocalReferenceResolver(@NotNull final PsiElement element) {
-        super(element);
+    private final PsiElement targetElement;
+
+    public LocalReferenceResolver(@NotNull final PsiElement souceElement, final PsiElement targetElement) {
+        super(souceElement);
+        this.targetElement = targetElement;
     }
 
     @Override
     public @NotNull ResolveResult @NotNull [] multiResolve(final boolean incompleteCode) {
-        return ofNullable(myElement.getUserData(ACTION_KEY)).flatMap(action -> {
-            final Project project = getProject(myElement);
-            return ofNullable(project)
-                    .flatMap(action::getLocalPath)
-                    .map(path -> LocalFileSystem.getInstance().findFileByPath(path))
-                    .map(virtualFile -> PsiManager.getInstance(project).findFile(virtualFile))
-                    .map(target -> target.getChildren().length > 0 ? target.getChildren()[0] : target)
-                    .map(target -> new ResolveResult[]{(new PsiElementResolveResult(target))});
-        }).orElse(ResolveResult.EMPTY_ARRAY);
+        if (targetElement != null) {
+            return new ResolveResult[]{new PsiElementResolveResult(targetElement)};
+        }
+        return ResolveResult.EMPTY_ARRAY;
     }
 
     @Nullable
@@ -43,6 +34,6 @@ public class LocalReferenceResolver extends PsiReferenceBase<PsiElement> impleme
 
     @Override
     public @NotNull String getCanonicalText() {
-        return myElement.getText();
+        return targetElement.getText();
     }
 }